Move CodeBlock into misc
