--- src/renderer/components/Settings/CodeBlock.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-
-import { isDevelopment } from "@/lib/utils";
-
-export const CodeBlock: React.FC<{
-    values: any
-}> = (props) => {
-    if (isDevelopment) {
-        return (<pre>{(JSON as any).stringify(props.values, 0, 2)}</pre>);
-    }
-    return null;
-};
